requirements.txt
# Core
torch
torchvision
torchsummary
numpy 
scipy
matplotlib

# Image processing
Pillow
opencv-python
opencv-contrib-python
opencv-python-headless
albumentations
nibabel
pydicom

# Training
wandb
tqdm

# Utiltiies
pytest
pyyaml
click
